feat : 채팅 메시지 말풍선 나누기

+++ android/Signal/app/src/main/java/com/ongo/signal/data/repository/chat/chatservice/ChatRepository.kt
@@ -10,6 +10,6 @@ interface ChatRepository {
 
     suspend fun saveChatRoom(chatRoom: ChatHomeCreate) : Response<ChatHomeCreate>
 
-    suspend fun getAllMessages(chat_id: Int): Response<MutableList<ChatHomeChildDto>>
+    suspend fun getAllMessages(chat_id: Long): Response<MutableList<ChatHomeChildDto>>
 
 }
